+++ src/Stormancer.Plugins/Users/Stormancer.Server.Plugins.Users/UserService.cs
@@ -20,7 +20,8 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-using Microsoft.EntityFrameworkCore;
+
+using Autofac.Core;
 using Nest;
 using Newtonsoft.Json.Linq;
 using Server.Plugins.Users;
@@ -30,7 +31,7 @@ using Stormancer.Server.Components;
 using Stormancer.Server.Plugins.Analytics;
 using Stormancer.Server.Plugins.Configuration;
 using Stormancer.Server.Plugins.Database;
-using Stormancer.Server.Plugins.Database.EntityFrameworkCore;
+
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -67,7 +68,7 @@ namespace Stormancer.Server.Plugins.Users
     {
 
         private readonly string _indexName;
-        private readonly DbContextAccessor _dbContext;
+        private readonly IUserStorage? _storage;
         private readonly ILogger _logger;
         private readonly Func<IEnumerable<IUserEventHandler>> _handlers;
         private readonly IConfiguration _configuration;
@@ -79,7 +80,7 @@ namespace Stormancer.Server.Plugins.Users
 
 
         public UserService(
-            DbContextAccessor dbContext,
+            IUserStorage? storage,
             IEnvironment environment,
             ILogger logger,
             Func<IEnumerable<IUserEventHandler>> eventHandlers,
@@ -91,7 +92,7 @@ namespace Stormancer.Server.Plugins.Users
             _handlers = eventHandlers;
             _configuration = configuration;
             _analytics = analytics;
-            _dbContext = dbContext;
+            _storage = storage;
             _logger = logger;
             //_logger.Log(LogLevel.Trace, "users", $"Using index {_indexName}", new { index = _indexName });
 
@@ -101,62 +102,29 @@ namespace Stormancer.Server.Plugins.Users
 
         public async Task<User> AddAuthentication(User user, string provider, string identifier, Action<dynamic> authDataModifier)
         {
-            var c = await _dbContext.GetDbContextAsync();
-
-            var auth = user.Auth[provider];
-            if (auth == null)
+            bool added = false;
+            if (_storage != null)
             {
-                auth = new JObject();
-                user.Auth[provider] = auth;
+                (user, added) = await _storage.GetAuthentication(user, provider, identifier, authDataModifier);
             }
-            authDataModifier?.Invoke(auth);
-
-            var userId = Guid.Parse(user.Id);
-            var record = c.Set<UserRecord>().Include(u => u.Identities.Where(i => i.Provider == provider && i.Identity == identifier)).SingleOrDefault(u => u.Id == userId);
 
 
-            var identity = record.Identities.FirstOrDefault();
-            var added = identity == null;
-            if (identity == null)
-            {
-                identity = new IdentityRecord { Provider = provider, Identity = identifier };
-                identity.Users = new List<UserRecord>
-                {
-                    record
-                };
 
-                await c.Set<IdentityRecord>().AddAsync(identity);
-            }
-            identity.MainUser = record;
-            record.Auth = JsonDocument.Parse(user.Auth.ToString());
-
-
-            await c.SaveChangesAsync();
-
-            var ctx = new AuthenticationChangedCtx(added ? AuthenticationChangedCtx.AuthenticationUpdateType.Add : AuthenticationChangedCtx.AuthenticationUpdateType.Update, provider, user);
-            await _handlers().RunEventHandler(h => h.OnAuthenticationChanged(ctx), ex => _logger.Log(LogLevel.Error, "user.addAuth", "An error occured while running OnAuthenticationChanged event handler", ex));
+            var ctx = new AuthenticationChangedCtx(_storage == null ? AuthenticationChangedCtx.AuthenticationUpdateType.None : added ? AuthenticationChangedCtx.AuthenticationUpdateType.Add : AuthenticationChangedCtx.AuthenticationUpdateType.Update, provider, user);
+            await _handlers().RunEventHandler(h => h.OnAuthenticationChanged(ctx), ex => _logger.Log(LogLevel.Error, "user.addAuth", "An error occurred while running OnAuthenticationChanged event handler", ex));
 
             return user;
         }
 
         public async Task<User> RemoveAuthentication(User user, string provider)
         {
-            var c = await _dbContext.GetDbContextAsync();
-
-            user.Auth.Remove(provider);
-            var userId = Guid.Parse(user.Id);
-            var record = c.Set<UserRecord>().Include(u => u.Identities.Where(i => i.Provider == provider)).SingleOrDefault(u => u.Id == userId);
-
-            if (record.Identities.Any())
+            if (_storage != null)
             {
-                c.Set<IdentityRecord>().RemoveRange(record.Identities);
+                user = await _storage.RemoveAuthentication(user, provider);
             }
 
-            record.Auth = JsonDocument.Parse(user.Auth.ToString());
 
-            await c.SaveChangesAsync();
-
-            var ctx = new AuthenticationChangedCtx(AuthenticationChangedCtx.AuthenticationUpdateType.Remove, provider, user);
+            var ctx = new AuthenticationChangedCtx(_storage == null ? AuthenticationChangedCtx.AuthenticationUpdateType.None : AuthenticationChangedCtx.AuthenticationUpdateType.Remove, provider, user);
             await _handlers().RunEventHandler(h => h.OnAuthenticationChanged(ctx), ex => _logger.Log(LogLevel.Error, "user.removeAuth", "An error occured while running OnAuthenticationChanged event handler", ex));
 
             return user;
@@ -165,28 +133,24 @@ namespace Stormancer.Server.Plugins.Users
         public async Task<User> CreateUser(string userId, JObject userData, string currentPlatform = "")
         {
             var user = new User() { Id = userId, LastPlatform = currentPlatform, UserData = userData };
-            var record = UserRecord.CreateRecordFromUser(user);
+            if (_storage != null)
+            {
+                user = await _storage.CreateUser(user);
+            }
 
-            var dbContext = await this._dbContext.GetDbContextAsync();
-            await dbContext.Set<UserRecord>().AddAsync(record);
 
-            await dbContext.SaveChangesAsync();
             _analytics.Push("user", "create", JObject.FromObject(new { UserId = user.Id, Platform = currentPlatform, createdOn = DateTime.UtcNow }));
             return user;
         }
 
         public async Task<string?> UpdateUserHandleAsync(string userId, string newHandle, CancellationToken cancellationToken)
         {
-            string GenerateUserHandleWithHash(string handle, UserHandleConfigSection section)
+            var section = _configuration.GetValue(UserHandleConfigSection.PATH, new UserHandleConfigSection());
+            string GenerateUserHandleWithHash(string handle)
             {
                 var suffix = _random.Next(0, (int)Math.Pow(10, section.HashDigits));
                 return handle + "#" + suffix;
             }
-
-
-
-            var section = _configuration.GetValue(UserHandleConfigSection.PATH, new UserHandleConfigSection());
-
 
             // Check handle validity
             if (!Regex.IsMatch(newHandle, section.Pattern))
@@ -198,57 +162,24 @@ namespace Stormancer.Server.Plugins.Users
                 throw new ClientException($"badHandle?tooLong&maxLength={section.MaxLength}");
             }
 
-            var ctx = new UpdateUserHandleCtx(userId, newHandle);
-            await _handlers().RunEventHandler(handler => handler.OnUpdatingUserHandle(ctx), ex => _logger.Log(LogLevel.Error, "usersessions", "An exception was thrown by an OnUpdatingUserHandle event handler", ex));
-
-            if (!ctx.Accept)
-            {
-                throw new ClientException(ctx.ErrorMessage);
-            }
-            var dbContext = await _dbContext.GetDbContextAsync();
-
-            var guid = Guid.Parse(userId);
-            var record = await dbContext.Set<UserRecord>().SingleOrDefaultAsync(u => u.Id == guid);
-
-            var userData = JObject.Parse(record.UserData.RootElement.GetRawText()!);
-            if (userData.TryGetValue(EphemeralAuthenticationProvider.IsEphemeralKey, out var isEphemeral) && (bool)isEphemeral)
+            if (_storage == null)
             {
                 throw new NotSupportedException();
-                //await UpdateHandleEphemeral();
-            }
-
-            if (section.AppendHash)
-            {
-                int tries = 0;
-                bool success = false;
-                do
-                {
-                    try
-                    {
-                        record.UserHandle = GenerateUserHandleWithHash(newHandle, section);
-                        await dbContext.SaveChangesAsync();
-                        success = true;
-                    }
-                    catch (Exception)
-                    {
-                        if (tries > 5)
-                        {
-                            throw;
-                        }
-                        tries++;
-
-                    }
-                }
-                while (!success);
             }
             else
             {
-                record.UserHandle = newHandle;
-                await dbContext.SaveChangesAsync();
+
+                var ctx = new UpdateUserHandleCtx(userId, newHandle);
+                await _handlers().RunEventHandler(handler => handler.OnUpdatingUserHandle(ctx), ex => _logger.Log(LogLevel.Error, "usersessions", "An exception was thrown by an OnUpdatingUserHandle event handler", ex));
+
+                if (!ctx.Accept)
+                {
+                    throw new ClientException(ctx.ErrorMessage);
+                }
+
+                return await _storage.UpdateUserHandleAsync(userId, newHandle, section.AppendHash ? GenerateUserHandleWithHash : null, cancellationToken);
             }
 
-
-            return record.UserHandle;
         }
 
         public class UserLastLoginUpdate
@@ -263,17 +194,17 @@ namespace Stormancer.Server.Plugins.Users
 
 
 
-        public async Task UpdateLastPlatform(string userId, string lastPlatform)
+        public Task UpdateLastPlatform(string userId, string lastPlatform)
         {
-            var guid = Guid.Parse(userId);
-            var dbContext = await _dbContext.GetDbContextAsync();
-            var record = await dbContext.Set<UserRecord>().SingleOrDefaultAsync(u => u.Id == guid);
-            if (record != null)
+            if (_storage != null)
             {
-                record.LastLogin = DateTime.Now;
-                record.LastPlatform = lastPlatform;
-                await dbContext.SaveChangesAsync();
+                return _storage.UpdateLastPlatform(userId, lastPlatform);
             }
+            else
+            {
+                return Task.CompletedTask;
+            }
+
         }
 
         private class ClaimUser
@@ -290,154 +221,105 @@ namespace Stormancer.Server.Plugins.Users
         }
 
 
+        private static Task<Dictionary<string, User?>> EmptyDictionaryTask = Task.FromResult(new Dictionary<string, User?>());
 
-
-        public async Task<Dictionary<string, User?>> GetUsersByIdentity(string provider, IEnumerable<string> identifiers)
+        public Task<Dictionary<string, User?>> GetUsersByIdentity(string provider, IEnumerable<string> identifiers)
         {
-            var c = await _dbContext.GetDbContextAsync();
-
-            var ids = identifiers.ToList();
-
-            var records = await c.Set<UserRecord>().Where(u => u.Identities.Any(i => i.Provider == provider && ids.Contains(i.Identity))).Include(u => u.Identities).ToListAsync();
-
-            var result = new Dictionary<string, User?>();
-            foreach (var id in ids)
+            if (_storage != null)
             {
-                var record = records.FirstOrDefault(r => r.Identities.Any(r => r.Identity == id && r.Provider == provider));
-                result[id] = UserRecord.CreateUserFromRecord(record);
+                return _storage.GetUsersByIdentity(provider, identifiers);
+            }
+            else
+            {
+                return EmptyDictionaryTask;
             }
 
-            return result;
         }
 
         public async Task<User?> GetUserByIdentity(string provider, string login)
         {
             var users = await GetUsersByIdentity(provider, Enumerable.Repeat(login, 1));
 
-            return users[login];
+            return users.TryGetValue(login, out var user) ? user : null;
         }
 
 
 
-        public async Task<User?> GetUser(string uid)
+        public Task<User?> GetUser(string uid)
         {
-            var dbContext = await _dbContext.GetDbContextAsync();
-
-            var guid = Guid.Parse(uid);
-            return UserRecord.CreateUserFromRecord(await dbContext.Set<UserRecord>().SingleOrDefaultAsync(u => u.Id == guid));
-
-        }
-
-        public async Task UpdateUserData<T>(string uid, T data)
-        {
-            var dbContext = await _dbContext.GetDbContextAsync();
-
-            var guid = Guid.Parse(uid);
-            var record = await dbContext.Set<UserRecord>().SingleOrDefaultAsync(u => u.Id == guid);
-            if (record == null)
+            if(_storage !=null)
             {
-                throw new InvalidOperationException($"notfound");
+                return _storage.GetUser(uid);
             }
             else
             {
-                if (data is JObject jObject)
-                {
-                    record.UserData = JsonDocument.Parse(jObject.ToString());
-                }
-                else
-                {
-                    record.UserData = JsonSerializer.SerializeToDocument(data!);
-                }
+                return Task.FromResult<User?>(null);
+            }
+           
 
+        }
 
-                await dbContext.SaveChangesAsync();
+        public Task UpdateUserData<T>(string uid, T data)
+        {
+           if(_storage != null)
+            {
+                return _storage.UpdateUserData<T>(uid, data);
+            }
+           else
+            {
+                return Task.CompletedTask;
             }
         }
-        public async Task<IEnumerable<User>> QueryUserHandlePrefix(string prefix, int take, int skip)
+        public Task<IEnumerable<User>> QueryUserHandlePrefix(string prefix, int take, int skip)
         {
-            var dbContext = await _dbContext.GetDbContextAsync();
-
-            var records = await dbContext.Set<UserRecord>().Where(u => u.UserHandle != null && u.UserHandle.StartsWith(prefix)).Skip(skip).Take(take).ToListAsync();
-
-            return records.Select(r => UserRecord.CreateUserFromRecord(r));
-
-        }
-        public async Task<IEnumerable<User>> Query(IEnumerable<KeyValuePair<string, string>> query, int take, int skip, CancellationToken cancellationToken)
-        {
-            var dbContext = await _dbContext.GetDbContextAsync();
-
-            var mainQuery = dbContext.Set<UserRecord>().AsQueryable().AsNoTracking();
-
-
-            foreach (var (key, value) in query)
+           if(_storage !=null)
             {
-                mainQuery = mainQuery.Where(r => r.UserData.RootElement.GetProperty("{" + key.Replace('.', ',') + "}").ToString() == value);
+                return _storage.QueryUserHandlePrefix(prefix, take, skip);
+            }
+           else
+            {
+                return Task.FromResult(Enumerable.Empty<User>());
             }
 
-            var results = await mainQuery.ToListAsync();
-
-            return results.Select(r => UserRecord.CreateUserFromRecord(r));
+        }
+        public Task<IEnumerable<User>> Query(IEnumerable<KeyValuePair<string, string>> query, int take, int skip, CancellationToken cancellationToken)
+        {
+            if(_storage != null)
+            {
+                return _storage.Query(query, take, skip, cancellationToken);
+            }
+           else
+            {
+                return Task.FromResult(Enumerable.Empty<User>());
+            }
         }
 
-        public async Task UpdateCommunicationChannel(string userId, string channel, JsonObject data)
-        {
-            var dbContext = await _dbContext.GetDbContextAsync();
 
-            var guid = Guid.Parse(userId);
-            var record = await dbContext.Set<UserRecord>().SingleOrDefaultAsync(u => u.Id == guid);
-            if (record == null)
+        public Task Delete(string id)
+        {
+            if(_storage != null)
             {
-                throw new InvalidOperationException($"notfound");
+                return _storage.Delete(id);
             }
             else
             {
-                var obj = JsonObject.Create(record.Channels.RootElement);
-                obj![channel] = data;
-
-                record.Channels = obj.Deserialize<JsonDocument>()!;
-                await dbContext.SaveChangesAsync();
+                return Task.CompletedTask;
             }
+           
 
         }
 
-        public async Task Delete(string id)
+        public Task<Dictionary<string, User?>> GetUsers(IEnumerable<string> userIds, CancellationToken cancellationToken)
         {
-            var dbContext = await _dbContext.GetDbContextAsync();
-
-            var guid = Guid.Parse(id);
-
-            var record = await dbContext.Set<UserRecord>().SingleOrDefaultAsync(u => u.Id == guid);
-
-            if (record != null)
+            if(_storage != null)
             {
-                dbContext.Set<UserRecord>().Remove(record);
-                await dbContext.SaveChangesAsync();
+                return _storage.GetUsers(userIds, cancellationToken);
+            }
+            else
+            {
+                return EmptyDictionaryTask;
             }
 
-        }
-
-        public async Task<Dictionary<string, User?>> GetUsers(IEnumerable<string> userIds, CancellationToken cancellationToken)
-        {
-
-            var dbContext = await _dbContext.GetDbContextAsync();
-
-
-            try
-            {
-                var guids = userIds.Select(id => Guid.TryParse(id, out var guid) ? guid : default).Where(guid => guid != Guid.Empty).ToArray();
-                var records = await dbContext.Set<UserRecord>().Where(u => guids.Contains(u.Id)).ToListAsync();
-                var results = new Dictionary<string, User?>();
-                foreach (var id in userIds)
-                {
-                    results[id] = UserRecord.CreateUserFromRecord(records.FirstOrDefault(r => r.Id == Guid.Parse(id)));
-                }
-                return results;
-            }
-            catch (FormatException ex)
-            {
-                _logger.Log(LogLevel.Error, "users", $"Failed to parse userIds ({string.Join(",", userIds)})", ex);
-                throw;
-            }
         }